Fix issue with table dependencies when dropping
Foreign key dependencies prevents dropping a table while there is still
some table which references the table to be dropped.

Since it was implicit that a dependency had to be declared before being
referenced, now it actually used by the system, by dropping tables in
the inverse order in which they are created.